Added light and dark themes to CSS code
I noticed a (presumably) unwanted behavior when the user had their browser set
to light mode - the text in the input area was dark, and almost
invisible.

I fixed it by adding css styles that incorporate the 'prefers-color-scheme'
media feature. This works by the browser determining which color theme
the user has requested (either in the browser's settings or system-wide
settings) and applying appropriate styles.

While I was at it, I decided to create a light mode that features new
colors. I'm not the best at color theory, so tweak it to your liking :)